requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
docstring-parser==0.17.0
fastapi==0.118.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.120.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.42.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
numpy==2.3.3
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.9
pydantic-core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tenacity==9.1.2
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
